Add (int) cast before _text_listeners.size() in jtextfield and jtextarea

+++ jgui/jtextarea.cpp
@@ -793,7 +793,7 @@ void TextArea::DispatchEvent(TextEvent *event)
 
 	int k=0;
 
-	while (k++ < _text_listeners.size()) {
+	while (k++ < (int)_text_listeners.size()) {
 		_text_listeners[k-1]->TextChanged(event);
 	}
